Rename searcher to searchers
