+++ src/smart_sauna_map/search_sauna.py
@@ -5,8 +5,8 @@ from functools import cache
 from typing import Optional
 
 from smart_sauna_map.data_models.sauna import Sauna
-from smart_sauna_map.searcher.abstract_searcher import AbstractSearcher
-from smart_sauna_map.searcher.sauna_ikitai_searcher import SaunaIkitaiSearcher
+from smart_sauna_map.searchers.abstract_searcher import AbstractSearcher
+from smart_sauna_map.searchers.sauna_ikitai_searcher import SaunaIkitaiSearcher
 
 __all__ = ["search_sauna"]
