+++ src/smart_sauna_map/searchers/abstract_searcher.py
