+++ src/smart_sauna_map/searchers/google_map_searcher.py
@@ -10,7 +10,7 @@ from requests.exceptions import HTTPError
 
 from smart_sauna_map.data_models.room import MansRoom, WomansRoom
 from smart_sauna_map.data_models.sauna import Sauna
-from smart_sauna_map.searcher.abstract_searcher import AbstractSearcher
+from smart_sauna_map.searchers.abstract_searcher import AbstractSearcher
 
 GOOGLE_MAP_API_KEY = os.environ.get("GOOGLE_MAP_API_KEY")
